requirements.txt
asgiref==3.3.4
beepy==1.0.7
Django==3.2
paho-mqtt==1.5.1
pytz==2021.1
simpleaudio==1.0.4
sqlparse==0.4.1
typing-extensions==3.7.4.3
